This application is a Java Spring Cloud application that act as a client in the [WEB-QUEUE-WORKER] architecture. This application connects to the [front-app] and passing request to the SQS. Also, recieves responses from SQS and send it to the [front-app]. The [front-app] will be communicated using [REST HTTP] and [Web Socket] protocols.
This application is a Java Spring Cloud application that act as a client in the [WEB-QUEUE-WORKER] architecture. This application connects to the [front-app] and passing request to the SQS. Also, recieves responses from SQS and send it to the [front-app]. The [front-app] will be communicated using [REST HTTP] and [Web Socket] protocols.